requirement.txt
Python 3.11.4
Flask 3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Flask-Migrate== 3.1.0